% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-loop.R
\name{begin_game}
\alias{begin_game}
\title{Start A 'r.oguelike' Game}
\usage{
begin_game()
}
\value{
Nothing. Clears the console and prints to it.
}
\description{
Clears the console and starts a game of 'r.oguelike'. The user inputs a
keypress to move the character and interact with the game.
}
\details{
Press W, A, S or D at the prompt and hit enter to move.
}
\examples{
\dontrun{begin_game()}
}
