% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dungeon.R
\name{generate_dungeon}
\alias{generate_dungeon}
\title{Generate A Dungeon Map}
\usage{
generate_dungeon(
  iterations = 5,
  n_row = 30,
  n_col = 50,
  n_rooms = 4,
  is_snake = FALSE,
  is_organic = TRUE,
  seed = NULL,
  colour = TRUE
)
}
\arguments{
\item{iterations}{Numeric.}

\item{n_row}{Numeric. Number of row tiles in the dungeon, i.e. its height.}

\item{n_col}{Numeric. Number of column tiles in the dungeon, i.e. its width.}

\item{n_rooms}{Numeric. Number of start points to be placed randomly to
'grow' rooms from.}

\item{is_snake}{Logical. Should the start points be connected in matrix index
order (\code{TRUE}), or randomly (\code{FALSE})? The former will create a
single continuous, winding cavern.}

\item{is_organic}{Logical. Join start points with corridors before
growing rooms (\code{TRUE}), or the other way round (\code{FALSE})? The
former looks more 'organic', the latter more square, or 'artificial'.}

\item{seed}{Numeric. Seed to reproduce a dungeon.}

\item{colour}{Logical. Should the characters be coloured using \pkg{crayon}
(\code{TRUE})?}
}
\value{
A matrix, invisibly. Output via \code{\link[base]{cat}} to the
console.
}
\description{
Generate a tile-based ASCII character dungeon map. Lays down 'start points'
and connecting corridors from which rooms can be grown.
}
\details{
You will have to experiment to find the 'best' argument values for
your needs. Typically, a larger dungeon should have more
\code{n_rooms} and can be grown through more \code{iterations}, and vice
versa.
}
\examples{
\dontrun{generate_dungeon(10, 20, 30, 3)}
}
