% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.oguelike-package.R
\docType{package}
\name{r.oguelike-package}
\alias{r.oguelike}
\alias{r.oguelike-package}
\title{r.oguelike: Play a Roguelike Game In The Console}
\description{
Play a game in the console that records your inputs to move a player around a generated world. Inspired heavily by roguelike games, particularly Rogue (1980) itself.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
