% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-movement.R
\name{.move_enemy}
\alias{.move_enemy}
\title{Move The Enemy To The Player}
\usage{
.move_enemy(room, dist)
}
\arguments{
\item{room}{Matrix. 2D room layout.}

\item{dist}{Matrix. Tile distance to player.}
}
\description{
Move The Enemy To The Player
}
