% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.oguelike-package.R
\docType{package}
\name{r.oguelike-package}
\alias{r.oguelike}
\alias{r.oguelike-package}
\title{r.oguelike: Play a Roguelike 'Game' in the Console}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Play a simple ASCII-only tile-based toy in the console, inspired heavily by roguelike games like Rogue (1980). Proof of concept.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/r.oguelike}
  \item \url{https://matt-dray.github.io/r.oguelike/}
  \item Report bugs at \url{https://github.com/matt-dray/r.oguelike/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
