% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.oguelike-package.R
\docType{package}
\name{r.oguelike-package}
\alias{r.oguelike}
\alias{r.oguelike-package}
\title{r.oguelike: Play a Roguelike Game In The Console}
\description{
Play a simple ASCII game in the console inspired heavily by roguelike games, particularly Rogue (1980) itself. Work in progress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/r.oguelike}
  \item \url{https://matt-dray.github.io/r.oguelike/}
  \item Report bugs at \url{https://github.com/matt-dray/r.oguelike/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
