% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roguelike.R
\name{start_game}
\alias{start_game}
\title{Start A 'r.oguelike' Game}
\usage{
start_game(max_turns = Inf)
}
\arguments{
\item{max_turns}{Numeric. How many turns? Default \code{Inf}.}
}
\value{
Nothing. Clears the console and prints to the console with
\code{cat}.
}
\description{
Clears the console and starts a game of 'r.oguelike'.
}
\details{
If your terminal supports the 'keypress' package, then you can use
single keypresses as input, otherwise you will have to type words at the
prompt. Use \code{\link[keypress]{has_keypress_support}} to see if
'keypress' is supported.
}
\examples{
\dontrun{start_game()}
}
